// packages/core/src/constants.d.ts
export declare const APP_INTERCEPTOR = "APP_INTERCEPTOR";
export declare const APP_PIPE = "APP_PIPE";
export declare const MESSAGES: {
    APPLICATION_START: string;
    APPLICATION_READY: string;
};
export declare const METADATA: {
    MODULES: string;
    IMPORTS: string;
    SERVICES: string;
    EXPORTS: string;
};
export declare const SHARED_MODULE_METADATA = "__sharedModule__";
export declare const GLOBAL_MODULE_METADATA = "__globalModule__";
export declare const PARAMTYPES_METADATA = "design:paramtypes";
export declare const SELF_DECLARED_DEPS_METADATA = "self:paramtypes";
export declare const OPTIONAL_DEPS_METADATA = "optional:paramtypes";
export declare const PROPERTY_DEPS_METADATA = "self:properties_metadata";
export declare const OPTIONAL_PROPERTY_DEPS_METADATA = "optional:properties_metadata";
export declare const PROVIDERS_METADATA = "__providers__";